Minor code fixes for possible build on Linux. Added directories with executable game for Windows and Linux

+++ Npc.cpp
@@ -104,8 +104,8 @@ bool Npc::isCollision(sf::Vector2f& playerPos)
 sf::Vector2f Npc::getNpcPosInTileGrid(sf::Vector2i& gridPos)
 {
 	// výpočet umístění npc na střed dlaždice mřížky
-	float centerX = std::floorf(((float)gridPos.x * MAP_DATA.tileSize) + (MAP_DATA.tileSize / 2));
-	float centerY = std::floorf(((float)gridPos.y * MAP_DATA.tileSize) + (MAP_DATA.tileSize / 2));
+	float centerX = std::floor(((float)gridPos.x * MAP_DATA.tileSize) + (MAP_DATA.tileSize / 2));
+	float centerY = std::floor(((float)gridPos.y * MAP_DATA.tileSize) + (MAP_DATA.tileSize / 2));
 
 	return sf::Vector2f(centerX, centerY);;
 }
